+++ pm4py/util/constants.py
@@ -1,19 +1,14 @@
-'''
-    This file is part of PM4Py (More Info: https://pm4py.fit.fraunhofer.de).
+import os
+import pkgutil
+from enum import Enum
 
-    PM4Py is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
 
-    PM4Py is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
+def get_param_from_env(name, default):
+    if name in os.environ and os.environ[name]:
+        return str(os.environ[name])
+    return default
 
-    You should have received a copy of the GNU General Public License
-    along with PM4Py.  If not, see <https://www.gnu.org/licenses/>.
-'''
+
 PARAMETER_CONSTANT_ACTIVITY_KEY = 'pm4py:param:activity_key'
 PARAMETER_CONSTANT_ATTRIBUTE_KEY = "pm4py:param:attribute_key"
 PARAMETER_CONSTANT_TIMESTAMP_KEY = 'pm4py:param:timestamp_key'
@@ -58,14 +53,15 @@ DEFAULT_ARTIFICIAL_END_ACTIVITY = "■"
 
 DEFAULT_BUSINESS_HOURS_WORKCALENDAR = None
 
-SHOW_EVENT_LOG_DEPRECATION = True
+SHOW_EVENT_LOG_DEPRECATION = True if get_param_from_env("PM4PY_SHOW_EVENT_LOG_DEPRECATION", "True").lower() == "true" else False
 TRIGGERED_DT_PARSING_WARNING = False
 
-DEFAULT_BGCOLOR = "white"
-DEFAULT_FORMAT_GVIZ_VIEW = "png"
+DEFAULT_BGCOLOR = get_param_from_env("PM4PY_DEFAULT_BGCOLOR", "white")
+DEFAULT_FORMAT_GVIZ_VIEW = get_param_from_env("PM4PY_DEFAULT_FORMAT_GVIZ_VIEW", "png")
 
-ENABLE_MULTIPROCESSING_DEFAULT = False
-DEFAULT_READ_XES_LEGACY_OBJECT = False
+ENABLE_MULTIPROCESSING_DEFAULT = True if get_param_from_env("PM4PY_ENABLE_MULTIPROCESSING_DEFAULT", "False").lower() == "true" else False
+DEFAULT_READ_XES_LEGACY_OBJECT = True if get_param_from_env("PM4PY_DEFAULT_READ_XES_LEGACY_OBJECT", "False").lower() == "true" else False
+DEFAULT_RETURN_DIAGNOSTICS_DATAFRAME = True if get_param_from_env("PM4PY_DEFAULT_RETURN_DIAGNOSTICS_DATAFRAME", "False").lower() == "true" else False
 
 # Default business hour slots: Mondays to Fridays, 7:00 - 17:00 (in seconds)
 DEFAULT_BUSINESS_HOUR_SLOTS = [
@@ -76,14 +72,12 @@ DEFAULT_BUSINESS_HOUR_SLOTS = [
     ((4 * 24 + 7) * 60 * 60, (4 * 24 + 17) * 60 * 60),
 ]
 
-OPENAI_MAX_LEN = 10000
-OPENAI_API_KEY = None
-OPENAI_DEFAULT_MODEL = "gpt-3.5-turbo"
-DEFAULT_GVIZ_VIEW = "view"
+OPENAI_MAX_LEN = int(get_param_from_env("PM4PY_OPENAI_MAX_LEN", "10000"))
+OPENAI_API_KEY = get_param_from_env("PM4PY_OPENAI_API_KEY", None)
+OPENAI_DEFAULT_MODEL = get_param_from_env("PM4PY_OPENAI_DEFAULT_MODEL", "gpt-3.5-turbo")
+DEFAULT_GVIZ_VIEW = get_param_from_env("PM4PY_DEFAULT_GVIZ_VIEW", None)
 
-import pkgutil
 if pkgutil.find_loader("psutil"):
-    import os
     import psutil
 
     parent_pid = os.getppid()
@@ -92,7 +86,9 @@ if pkgutil.find_loader("psutil"):
     if "PBIDesktop" in parent_name:
         DEFAULT_GVIZ_VIEW = "matplotlib_view"
 
-from enum import Enum
+
+if DEFAULT_GVIZ_VIEW is None:
+    DEFAULT_GVIZ_VIEW = "view"
 
 
 class AvailableSerializations(Enum):