+++ pm4py/visualization/sna/variants/pyvis.py
@@ -91,8 +91,13 @@ def apply(sna: SNA, parameters=None):
     got_net.show_buttons(filter_=['nodes', 'edges', 'physics'])
 
     F = open(temp_file_name, "w")
-    F.write(got_net.generate_html())
-    F.close()
+    try:
+        F.write(got_net.generate_html())
+        F.close()
+    except:
+        # networkx 3.1
+        F.close()
+        got_net.write_html(temp_file_name)
     
     return temp_file_name