+++ CHANGELOG.md
@@ -13,6 +13,8 @@
   * improved integration possibilities with PowerBI
 * f805fd46673be291584ce489bf9def73df1dba71
   * read constants from environment variables
+* 6737019a6de4e15d6063506e5a7ea2e571fc167d
+  * added DEFAULT_RETURN_DIAGNOSTICS_DATAFRAME option in constants
 
 ### Deprecated